Refactor seed.js Split into two files. One file (seedHelpers.js) contains the seeding-related functions. The other file (seed.js) automatically seeds the database when it it run.
